- em modo demo, na tela de seleção, está entrando na fase do zodiac ao invés de na fase snow - slide/dash não funciona na água - teleport out sem transparência (fundo preto) - novo tipo de teleportadores que setam scroll para lock à esquerda ou direita após a troca da mapa
- Rockbot1:
    # projétil do gambit está muito lento
    # morcego tem muitos pontos de vida
    # faltando tiles na sala de chefe do seahorse
    # alguns fundos de mapa não funciona corretamente (é preciso ser 320px de largura agora)
    # castle5: ao ir para teleportadores, fica com scroll-x errado (mais para a direita do que deveria)

+++ character/artificial_inteligence.cpp
@@ -260,7 +260,7 @@ void artificial_inteligence::execute_ai_step()
         execute_ai_wait_random_time();
     } else {
         std::cout << "AI_ACTION_JUMP_ATTACK_UP: " << (int)AI_ACTION_JUMP_ATTACK_UP << std::endl;
-        std::cout << "********** AI number[" << _number << "], pos[" << _ai_chain_n << "], _current_ai_type[" << _current_ai_type << "] - NOT IMPLEMENTED *******" << std::endl;
+        std::cout << "********** AI number[" << _number << "], pos[" << _ai_chain_n << "], _current_ai_type[" << (int)_current_ai_type << "] - NOT IMPLEMENTED *******" << std::endl;
     }
 }
